Refactor bit handling and add new field classes
Refactored bit handling across multiple classes, changing `BitsCount` properties to `byte?` and updating methods to return `byte` instead of `int`. Removed `availableTypes` array and `MapSymbolToSpecialType` method from `AdvancedBitFieldPropertyAggregator`, and refactored `Aggregate` method. Updated `BaseBitFieldPropertyAggregator` with new methods for property grouping, bit distribution, size validation, and type mapping. Enhanced `DictionariesPool<TKey, TValue>` and `ListsPool<T>` to pre-allocate instances and limit pool size. Added `AddRange(IEnumerable<T> items)` to `ImmutableArrayBuilder<T>`. Changed `ExistingSymbol` property type in `FieldName` and `IFieldName` to `IFieldSymbol?`. Introduced new classes: `FieldNameEqualityComparer`, `ExistingFieldAggregator`, `BitSize` enum, `ExistingFieldGsr`, and `ExistingFieldRequest`.

+++ PropertyBitPack.SourceGen/BitCountHelper.cs
@@ -19,7 +19,7 @@ internal static class BitCountHelper
     /// </summary>
     /// <param name="property">The bit field property info.</param>
     /// <returns>The number of bits required.</returns>
-    public static int GetEffectiveBitsCount(BaseBitFieldPropertyInfo property)
+    public static byte GetEffectiveBitsCount(BaseBitFieldPropertyInfo property)
     {
         // If the attribute explicitly sets BitsCount, return that value.
         if (property.AttributeParsedResult.BitsCount.HasValue)
@@ -33,7 +33,7 @@ internal static class BitCountHelper
         return GetBitsCountForSpecialType(property.PropertyType.SpecialType); ;
     }
 
-    public static int GetBitsCountForSpecialType(SpecialType specialType)
+    public static byte GetBitsCountForSpecialType(SpecialType specialType)
     {
         return specialType switch
         {
@@ -54,7 +54,7 @@ internal static class BitCountHelper
     /// (e.g. System_Byte=8, System_UInt16=16, etc.).
     /// Returns 0 if not recognized.
     /// </summary>
-    public static int GetTypeBitCapacity(SpecialType st)
+    public static byte GetTypeBitCapacity(SpecialType st)
     {
         return st switch
         {
